+++ local_planner/launch/local_planner_stereo.launch
@@ -23,6 +23,7 @@
     <param name="goal_y_param" value="15"/>
     <param name="goal_z_param" value="4" />
     <param name="depth_points_topic" value="/stereo/points2" />
+    <param name="world_name" value="$(find local_planner)/../sim/worlds/simple_obstacle.yaml" />
   </node>
 
   <node name="rviz" pkg="rviz" type="rviz" output="screen" args="-d $(find local_planner)/resource/local_planner.rviz" />
