+++ local_planner/src/nodes/local_planner_node.cpp
@@ -931,13 +931,13 @@ int main(int argc, char **argv) {
     hover = false;
 
     //visualize world in RVIZ
-      if(!Node.world_path_.empty() && startup){
-      	  visualization_msgs::MarkerArray marker_array;
-      	  if(!visualizeRVIZWorld(marker_array, Node.world_path_)){
-      	     Node.world_pub_.publish(marker_array);
-      	  }
-      	  startup = false;
+    if(!Node.world_path_.empty() && startup){
+      visualization_msgs::MarkerArray marker_array;
+      if(!visualizeRVIZWorld(marker_array, Node.world_path_)){
+        Node.world_pub_.publish(marker_array);
       }
+      startup = false;
+    }
 
     // Process callbacks & wait for a position update
     while (!Node.position_received_ && ros::ok()) {
