+++ local_planner/src/nodes/rviz_world_loader.h
@@ -1,5 +1,5 @@
-#ifndef RVIZ_LOAD_H
-#define RVIZ_LOAD_H
+#ifndef RVIZ_WORLD_H
+#define RVIZ_WORLD_H
 
 
 #include "yaml-cpp/yaml.h"
@@ -7,6 +7,7 @@
 #include <fstream>
 #include <string>
 #include <vector>
+#include <Eigen/Core>
 
 #include <ros/ros.h>
 #include <visualization_msgs/Marker.h>
@@ -15,27 +16,19 @@
 #include <sys/stat.h>
 
 //data types
-struct Vec3 {
-   float x, y, z;
-};
-
-struct Vec4 {
-   float x, y, z, w;
-};
-
 struct world_object{
    std::string type;
    std::string name;
    std::string frame_id;
    std::string mesh_resource;
-   Vec3 position;
-   Vec4 orientation;
-   Vec3 scale;
+   Eigen::Vector3f position;
+   Eigen::Vector4f orientation;
+   Eigen::Vector3f scale;
 };
 
-//extraction operators for these types
-void operator >> (const YAML::Node& node, Vec3& v);
-void operator >> (const YAML::Node& node, Vec4& v);
+//extraction operators
+void operator >> (const YAML::Node& node, Eigen::Vector3f& v);
+void operator >> (const YAML::Node& node, Eigen::Vector4f& v);
 void operator >> (const YAML::Node& node, world_object& item);
 
 //helper function to resolve gazebo model paths
@@ -44,4 +37,4 @@ int resolveUri(std::string& uri);
 //yaml parsing and marker publishing
 int visualizeRVIZWorld(visualization_msgs::MarkerArray& marker_array, std::string world_path);
 
-#endif  // RVIZ_LOAD_H
+#endif  // RVIZ_WORLD_H
